+++ package.json
@@ -38,6 +38,7 @@
     "react-bootstrap": "^0.31.5",
     "react-dev-utils": "^4.0.0",
     "react-dom": "^15.6.1",
+    "react-router-dom": "^4.2.2",
     "react-tabs": "^2.1.1",
     "style-loader": "0.18.2",
     "sw-precache-webpack-plugin": "0.11.4",
