+++ src/styles/blocks/tabs.scss
@@ -2,7 +2,7 @@
   display: block;
   width: 100vw;
   height: 100%;
-  margin: 0 auto;
+  margin: 40px auto;
 }
 
 .devices {
